// SmartTooling-SPA/src/app/_core/_models/best-line/critical-process-report-dto.ts
export interface CriticalProcessReportDTO {
    factory_id: string;
    line_id: string;
    line_name: string;
    line_type_id: string;
    line_type_name: string;
    model_no: string;
    model_name: string;
    stage_id: string;
    stage_name: string;
    operation_id: string;
    operation_name_zh: string;
    operation_name_en: string;
    operation_name_local: string;
}